gcp/requirements.txt
google-cloud-storage == 1.16.1
absl-py==0.15.0
anyio==3.6.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
charset-normalizer==2.1.1
clang==5.0
click==8.1.3
colorama==0.4.5
fastapi==0.80.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h11==0.13.0
h5py==3.1.0
idna==3.3
importlib-metadata==4.12.0
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==9.1.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
python-multipart==0.0.5
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.15.0
sniffio==1.2.0
starlette==0.19.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.4
tensorflow-estimator==2.6.0
termcolor==1.1.0
typing-extensions==3.10.0.2
urllib3==1.26.12
uvicorn==0.18.3
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.8.1



